+++ examples/scripts/spm_nlopt.py
@@ -6,11 +6,11 @@ Measurements = pd.read_csv("examples/scripts/Chen_example.csv", comment="#").to_
 dataset = [
     pybop.Dataset("Time [s]", Measurements[:, 0]),
     pybop.Dataset("Current function [A]", Measurements[:, 1]),
-    pybop.Dataset("Terminal voltage [V]", Measurements[:, 2]),
+    pybop.Dataset("Voltage [V]", Measurements[:, 2]),
 ]
 
 # Define model
-parameter_set = pybop.ParameterSet("pybamm", "Chen2020")
+parameter_set = pybop.ParameterSet.pybamm("Chen2020")
 model = pybop.models.lithium_ion.SPM(
     parameter_set=parameter_set, options={"thermal": "lumped"}
 )
@@ -30,7 +30,7 @@ parameters = [
 ]
 
 # Define the cost to optimise
-signal = "Terminal voltage [V]"
+signal = "Voltage [V]"
 problem = pybop.FittingProblem(model, parameters, dataset, signal=signal, init_soc=0.98)
 cost = pybop.RootMeanSquaredError(problem)